changed from parseint to parsefloat

+++ v0/destinations/iterable/transform.js
@@ -116,13 +116,13 @@ function constructPayload(message, category, destination) {
             ","
           );
         }
-        rawPayload.items[i].price = parseInt(rawPayload.items[i].price);
+        rawPayload.items[i].price = parseFloat(rawPayload.items[i].price);
         rawPayload.items[i].quantity = parseInt(rawPayload.items[i].quantity);
       }
       rawPayload.createdAt = parseInt(
         new Date(rawPayload.createdAt).getTime() / 1000
       );
-      rawPayload.total = parseInt(rawPayload.total);
+      rawPayload.total = parseFloat(rawPayload.total);
       if (rawPayload.campaignId)
         rawPayload.campaignId = parseInt(rawPayload.campaignId);
       if (rawPayload.templateId)
@@ -143,7 +143,7 @@ function constructPayload(message, category, destination) {
             ","
           );
         }
-        rawPayload.items[i].price = parseInt(rawPayload.items[i].price);
+        rawPayload.items[i].price = parseFloat(rawPayload.items[i].price);
         rawPayload.items[i].quantity = parseInt(rawPayload.items[i].quantity);
       }
       break;
